% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_model.R
\name{constant_model}
\alias{constant_model}
\title{constant model}
\usage{
constant_model(in_dat, year_begin, year_end)
}
\arguments{
\item{year_begin}{begin of prediction}

\item{year_end}{end of prediction}

\item{points_dat}{tibble with model results of previous period by spatial_unit, nat}
}
\value{
#tibble with prediction data
}
\description{
constant model
}
