% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_fertility_rate.R
\name{forecast_fertility_rate}
\alias{forecast_fertility_rate}
\title{Forecast age-specific fertility rates}
\usage{
forecast_fertility_rate(
  fer_dat,
  tfr_dat,
  mab_dat,
  year_begin,
  year_end,
  maxit = 1000,
  abstol = 0.001
)
}
\arguments{
\item{fer_dat}{fertility data, tibble with variables spatial_unit, nat, age, fer}

\item{tfr_dat}{tfr data, tibble with variables spatial_unit, nat, year, tfr}

\item{mab_dat}{mab data, tibble with variables spatial_unit, nat, year, mab}

\item{year_begin}{forecast begin}

\item{year_end}{forecast end}

\item{maxit}{maximum iterations of optimization}

\item{abstol}{absolute tolerance of optimization}
}
\value{
birth rate, tibble with variables spatial_unit, nat, age, birth_rate
}
\description{
Forecast age-specific fertility rates
}
