% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_tfr_mab.R
\name{forecast_tfr_mab}
\alias{forecast_tfr_mab}
\title{forecast tfr or mab}
\usage{
forecast_tfr_mab(
  topic,
  input_dataset,
  trend_model,
  temporal_model,
  temporal_end = NA,
  constant_model
)
}
\arguments{
\item{topic}{tfr or mab}

\item{input_dataset}{input tibble with variables: spatial_unit, nat, year, y (tfr or mab)}

\item{trend_model}{model type (ARIMA or lm), begin, end, amount of years used to fit the lm model}

\item{temporal_model}{model type (cubic, Bezier or constant), begin, end,
proportion of trend used, proportion slopes used,}

\item{temporal_end}{tibble with given y-values at the end of the temporal forecast period}

\item{constant_model}{model type (constant), begin, end}
}
\value{
predictions (tibble)
}
\description{
forecast tfr or mab
}
