% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_birth.R
\docType{data}
\name{fso_birth}
\alias{fso_birth}
\title{FSO data for births}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 437054 rows and 5 columns.
}
\usage{
fso_birth
}
\description{
Data was retrieved from the FSO.
}
\keyword{datasets}
