% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_pop.R
\docType{data}
\name{fso_pop}
\alias{fso_pop}
\title{Population data from the Federal Statistical Office}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2940 rows and 5 columns.
}
\usage{
fso_pop
}
\description{
End year population from different spatial units.
}
\keyword{datasets}
