% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birth_data.R
\name{get_birth_data}
\alias{get_birth_data}
\title{FSO births}
\usage{
get_birth_data(
  year_first,
  year_last,
  age_fert_min,
  age_fert_max,
  spatial_code,
  spatial_unit,
  with_nationality = TRUE
)
}
\arguments{
\item{year_first}{first year}

\item{year_last}{last year}

\item{age_fert_min}{minimum age (of 'fertile age')}

\item{age_fert_max}{maximum age (of 'fertile age')}

\item{spatial_code}{official FSO codes, vector}

\item{spatial_unit}{spatial unit names (free choice), vector}

\item{with_nationality}{TRUE (with nationality), FALSE (without nationality)}
}
\value{
births, tibble (spatial_unit, year, nationality, age, pop),
with or without nat (nationality)
}
\description{
FSO births
}
