% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{get_input_data}
\alias{get_input_data}
\title{input data}
\usage{
get_input_data(
  pop,
  bir,
  year_first,
  year_last,
  age_fert_min,
  age_fert_max,
  fer_last_years
)
}
\arguments{
\item{pop}{female population at 'fertile age',

tibble (spatial_unit, year, age, pop), with or without nat (nationality)}

\item{bir}{births

tibble (spatial_unit, year, age, pop), with or without nat (nationality)}

\item{year_first}{first year}

\item{year_last}{last year}

\item{age_fert_min}{minimum age (of 'fertile age')}

\item{age_fert_max}{maximum age (of 'fertile age')}

\item{fer_last_years}{how many years are used to calculate age-specific fertility rates?}
}
\value{
list with:
tfr (total fertility rate),
mab (mean age of the mother at birth),
fer (fertility rate of last year(s))
}
\description{
input data
}
