% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_tibble_with_cols}
\alias{is_tibble_with_cols}
\title{Check: tibble with columns spatial_unit and nat?}
\usage{
is_tibble_with_cols(x)
}
\arguments{
\item{x}{tibble}
}
\value{
TRUE if tibble with these columns, else FALSE
}
\description{
Check: tibble with columns spatial_unit and nat?
}
