% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab_fun.R
\name{mab_fun}
\alias{mab_fun}
\title{mab (mean age at birth) function}
\usage{
mab_fun(a0, age, y_no_a0, mab_proj)
}
\arguments{
\item{a0}{intercept (to be optimized), one value}

\item{age}{age (from age_min - 1 until age_max), vector}

\item{y_no_a0}{regression result without intercept, vector}

\item{mab_proj}{projected mab (objective), one value}
}
\value{
diff (absolute difference between mab and mab_proj), one value
}
\description{
mab (mean age at birth) function
}
