% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_fun.R
\name{opt_fun}
\alias{opt_fun}
\title{optimize the mab function}
\usage{
opt_fun(mab_proj, a0, age, y_no_a0, maxit, abstol)
}
\arguments{
\item{mab_proj}{objective of the optimization}

\item{a0}{intercept (to be optimized), one value}

\item{age}{age (from age_min - 1 until age_max), vector}

\item{y_no_a0}{regression result without intercept, vector}

\item{maxit}{maximum iterations (of optim function)}

\item{abstol}{absolute tolerance (of optim function)}
}
\value{
tibble with optimized parameter and objective function
}
\description{
optimize the mab function
}
