% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_regression.R
\name{pred_regression}
\alias{pred_regression}
\title{regression model: fit and prediction}
\usage{
pred_regression(y, x1, x2, x3, x4, x5, age_min, age_max)
}
\arguments{
\item{y}{regression fit}

\item{x1}{explanatory variable (age + 0.5}

\item{x2}{explanatory variable (x1 ^ 2)}

\item{x3}{explanatory variable (x1 ^ 3)}

\item{x4}{explanatory variable (x1 ^ 4)}

\item{x5}{explanatory variable (x1 ^ 5)}

\item{age_min}{minimum fertile age}

\item{age_max}{maximum fertile age}
}
\description{
regression model: fit and prediction
}
