% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_Bezier.R
\name{temporal_Bezier}
\alias{temporal_Bezier}
\title{temporal model: Bézier}
\usage{
temporal_Bezier(points_dat, year_begin, year_end)
}
\arguments{
\item{points_dat}{tibble with point information (x0, y0, z0, x1, y1, z1) by spatial_unit, nat}

\item{year_begin}{begin of prediction}

\item{year_end}{end of prediction}
}
\value{
#tibble with prediction data
}
\description{
temporal model: Bézier
}
