% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_constant.R
\name{temporal_constant}
\alias{temporal_constant}
\title{temporal model: constant}
\usage{
temporal_constant(points_dat, year_begin, year_end)
}
\arguments{
\item{points_dat}{tibble with point information (x0, y0) by spatial_unit, nat}

\item{year_begin}{begin of prediction}

\item{year_end}{end of prediction}
}
\value{
#tibble with prediction data
}
\description{
temporal model: constant
}
