% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_points.R
\name{temporal_points}
\alias{temporal_points}
\title{temporal model: start and end point}
\usage{
temporal_points(
  input_past,
  input_trend,
  year_begin,
  year_end,
  trend_prop,
  z0_prop,
  z1_prop
)
}
\arguments{
\item{input_past}{#tibble with variables year, x, spatial_unit, nat}

\item{input_trend}{#tibble with variables year, x, spatial_unit, nat, category}

\item{year_begin}{begin of prediction (temporal model)}

\item{year_end}{end of prediction (temporal model)}

\item{trend_prop}{y value of the end point: proportion of trend vs. past}

\item{z0_prop}{proportion of the calculated slope of the start point (z0)}

\item{z1_prop}{proportion of the calculated slope of the end point (z1)}
}
\description{
temporal model: start and end point
}
