% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_arima.R
\name{trend_arima}
\alias{trend_arima}
\title{trend prediction with arima model}
\usage{
trend_arima(year, y, year_begin, year_end, trend_prop)
}
\arguments{
\item{year}{years of input data}

\item{y}{y values of input data (e.a. TFR or MAB)}

\item{year_begin}{begin of prediction}

\item{year_end}{end of prediction}

\item{trend_prop}{y value of the end point: proportion of trend vs. past}
}
\value{
tibble prediction data
}
\description{
trend prediction with arima model
}
