% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_lm.R
\name{trend_lm}
\alias{trend_lm}
\title{trend prediction with linear model}
\usage{
trend_lm(input_lm, year_begin, year_end, trend_past, trend_prop)
}
\arguments{
\item{input_lm}{tibble with variables year, y, spatial_unit, nat}

\item{year_begin}{begin of prediction}

\item{year_end}{end of prediction}

\item{trend_past}{amount of past years that are used to fit the model}

\item{trend_prop}{y value of the end point: proportion of trend vs. past}
}
\value{
tibble with prediction data
}
\description{
trend prediction with linear model
}
