% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_model.R
\name{constant_model}
\alias{constant_model}
\title{Constant model}
\usage{
constant_model(points_dat, year_begin, year_end)
}
\arguments{
\item{points_dat}{tibble with model results of the previous period by
\code{spatial_unit} and \code{nat}.}

\item{year_begin}{first year of the prediction.}

\item{year_end}{last year of the prediction.}
}
\value{
tibble with prediction data
}
\description{
Constant model
}
