% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_input_data.R
\name{create_input_data}
\alias{create_input_data}
\title{Create model input data}
\usage{
create_input_data(
  population,
  births,
  year_first,
  year_last,
  age_fert_min,
  age_fert_max,
  fer_last_years
)
}
\arguments{
\item{population}{data frame, female population at 'fertile age'; columns:
spatial_unit, year, age, pop, with or without nat (nationality).}

\item{births}{data frame, births of females in the fertile age range; columns:
spatial_unit, year, age, pop, with or without nat (nationality).}

\item{year_first}{numeric, first year.}

\item{year_last}{numeric, last year.}

\item{age_fert_min}{numeric, minimum age (of 'fertile age').}

\item{age_fert_max}{numeric, minimum age (of 'fertile age').}

\item{fert_hist_years}{how many years are used to calculate age-specific fertility rates?}
}
\value{
list with:
tfr (total fertility rate),
mab (mean age of the mother at birth),
fer (fertility rate of last year(s))
}
\description{
Create model input data
}
\examples{
make_input_data(
  population = fso_pop,
  births = fso_birth |> 
    dplyr::filter(spatial_unit \%in\% c("Stadt Zürich", "Frauenfeld", "Uster")),
  year_first = 2011,
  year_last = 2023,
  age_fert_min = 15,
  age_fert_max = 49,
  fer_last_years = 1
)
}
