% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fso_birth.R
\docType{data}
\name{fso_birth}
\alias{fso_birth}
\title{FSO data for births}
\format{
A data frame with 437054 rows and 5 variables:
\describe{
\item{year}{year within the range 2010-2023}
\item{spatial_unit}{Swiss municipalities by name}
\item{nat}{Nationality; either "ch" = Swiss or "int" = international}
\item{age}{Age in full years}
\item{n_birth}{number of births}
}
}
\usage{
fso_birth
}
\description{
Data was retrieved from the FSO.
}
\keyword{datasets}
