% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birth_data.R
\name{get_birth_data}
\alias{get_birth_data}
\title{Get FSO population (females at 'fertile' age)}
\usage{
get_birth_data()
}
\description{
Data was retrieved from the FSO.
}
\examples{
get_birth_data()
}
