% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_data.R
\name{get_population_data}
\alias{get_population_data}
\title{FSO population (females at 'fertile' age)}
\usage{
get_population_data(
  number_fso,
  year_first,
  year_last,
  age_fert_min,
  age_fert_max,
  spatial_code,
  spatial_unit,
  with_nationality = TRUE
)
}
\arguments{
\item{number_fso}{number of FSO table (STAT-TAB)}

\item{year_first}{first year}

\item{year_last}{last year}

\item{age_fert_min}{minimum age (of 'fertile age')}

\item{age_fert_max}{maximum age (of 'fertile age')}

\item{spatial_code}{official FSO codes, vector}

\item{spatial_unit}{spatial unit names (free choice), vector}

\item{with_nationality}{TRUE (with nationality), FALSE (without nationality)}
}
\value{
female population at 'fertile age' at the end of the year,
tibble (spatial_unit, year, age, pop), with or without nat (nationality)
}
\description{
FSO population (females at 'fertile' age)
}
